LayoutTests/webcl/tests/webcl-test-consts.html

<html>
<head>
<title>WebCL webcl.idl constants tests.</title>
<link rel="stylesheet" href="../resources/js-test-style.css"/>
<script type="text/javascript" src="../resources/js-test-pre.js"></script>
<script type="text/javascript">

function runTest() {

    shouldBeTrue("webcl.SUCCESS == 0");
    shouldBeTrue("webcl.DEVICE_NOT_FOUND == -1");
    shouldBeTrue("webcl.DEVICE_NOT_AVAILABLE == -2");
    shouldBeTrue("webcl.COMPILER_NOT_AVAILABLE == -3");
    shouldBeTrue("webcl.MEM_OBJECT_ALLOCATION_FAILURE == -4");
    shouldBeTrue("webcl.OUT_OF_RESOURCES == -5");
    shouldBeTrue("webcl.OUT_OF_HOST_MEMORY == -6");
    shouldBeTrue("webcl.PROFILING_INFO_NOT_AVAILABLE == -7");
    shouldBeTrue("webcl.MEM_COPY_OVERLAP == -8");
    shouldBeTrue("webcl.IMAGE_FORMAT_MISMATCH == -9");
    shouldBeTrue("webcl.IMAGE_FORMAT_NOT_SUPPORTED == -10");
    shouldBeTrue("webcl.BUILD_PROGRAM_FAILURE == -11");
    shouldBeTrue("webcl.MAP_FAILURE == -12");
    shouldBeTrue("webcl.MISALIGNED_SUB_BUFFER_OFFSET == -13");
    shouldBeTrue("webcl.EXEC_STATUS_ERROR_FOR_EVENTS_IN_WAIT_LIST == -14");
    shouldBeTrue("webcl.INVALID_VALUE == -30");
    shouldBeTrue("webcl.INVALID_DEVICE_TYPE == -31");
    shouldBeTrue("webcl.INVALID_PLATFORM == -32");
    shouldBeTrue("webcl.INVALID_DEVICE == -33");
    shouldBeTrue("webcl.INVALID_CONTEXT == -34");
    shouldBeTrue("webcl.INVALID_QUEUE_PROPERTIES == -35");
    shouldBeTrue("webcl.INVALID_COMMAND_QUEUE == -36");
    shouldBeTrue("webcl.INVALID_HOST_PTR == -37");
    shouldBeTrue("webcl.INVALID_MEM_OBJECT == -38");
    shouldBeTrue("webcl.INVALID_IMAGE_FORMAT_DESCRIPTOR == -39");
    shouldBeTrue("webcl.INVALID_IMAGE_SIZE == -40");
    shouldBeTrue("webcl.INVALID_SAMPLER == -41");
    shouldBeTrue("webcl.INVALID_BINARY == -42");
    shouldBeTrue("webcl.INVALID_BUILD_OPTIONS == -43");
    shouldBeTrue("webcl.INVALID_PROGRAM == -44");
    shouldBeTrue("webcl.INVALID_PROGRAM_EXECUTABLE == -45");
    shouldBeTrue("webcl.INVALID_KERNEL_NAME == -46");
    shouldBeTrue("webcl.INVALID_KERNEL_DEFINITION == -47");
    shouldBeTrue("webcl.INVALID_KERNEL == -48");
    shouldBeTrue("webcl.INVALID_ARG_INDEX == -49");
    shouldBeTrue("webcl.INVALID_ARG_VALUE == -50");
    shouldBeTrue("webcl.INVALID_ARG_SIZE == -51");
    shouldBeTrue("webcl.INVALID_KERNEL_ARGS == -52");
    shouldBeTrue("webcl.INVALID_WORK_DIMENSION == -53");
    shouldBeTrue("webcl.INVALID_WORK_GROUP_SIZE == -54");
    shouldBeTrue("webcl.INVALID_WORK_ITEM_SIZE == -55");
    shouldBeTrue("webcl.INVALID_GLOBAL_OFFSET == -56");
    shouldBeTrue("webcl.INVALID_EVENT_WAIT_LIST == -57");
    shouldBeTrue("webcl.INVALID_EVENT == -58");
    shouldBeTrue("webcl.INVALID_OPERATION == -59");
    shouldBeTrue("webcl.INVALID_GL_OBJECT == -60");
    shouldBeTrue("webcl.INVALID_BUFFER_SIZE == -61");
    shouldBeTrue("webcl.INVALID_MIP_LEVEL == -62");
    shouldBeTrue("webcl.INVALID_GLOBAL_WORK_SIZE == -63");
    shouldBeTrue("webcl.INVALID_PROPERTY == -64");

    shouldBeTrue("webcl.PLATFORM_PROFILE == 0x0900");
    shouldBeTrue("webcl.PLATFORM_VERSION == 0x0901");

    shouldBeTrue("webcl.DEVICE_TYPE_DEFAULT == 0x1");
    shouldBeTrue("webcl.DEVICE_TYPE_CPU == 0x2");
    shouldBeTrue("webcl.DEVICE_TYPE_GPU == 0x4");
    shouldBeTrue("webcl.DEVICE_TYPE_ACCELERATOR == 0x8");
    shouldBeTrue("webcl.DEVICE_TYPE_ALL == 0xFFFFFFFF");

    shouldBeTrue("webcl.DEVICE_TYPE == 0x1000");
    shouldBeTrue("webcl.DEVICE_MAX_COMPUTE_UNITS == 0x1002");
    shouldBeTrue("webcl.DEVICE_MAX_WORK_ITEM_DIMENSIONS == 0x1003");
    shouldBeTrue("webcl.DEVICE_MAX_WORK_GROUP_SIZE == 0x1004");
    shouldBeTrue("webcl.DEVICE_MAX_WORK_ITEM_SIZES == 0x1005");
    shouldBeTrue("webcl.DEVICE_PREFERRED_VECTOR_WIDTH_CHAR == 0x1006");
    shouldBeTrue("webcl.DEVICE_PREFERRED_VECTOR_WIDTH_SHORT == 0x1007");
    shouldBeTrue("webcl.DEVICE_PREFERRED_VECTOR_WIDTH_INT == 0x1008");
    shouldBeTrue("webcl.DEVICE_PREFERRED_VECTOR_WIDTH_LONG == 0x1009");
    shouldBeTrue("webcl.DEVICE_PREFERRED_VECTOR_WIDTH_FLOAT == 0x100A");
    shouldBeTrue("webcl.DEVICE_PREFERRED_VECTOR_WIDTH_DOUBLE == 0x100B");
    shouldBeTrue("webcl.DEVICE_ADDRESS_BITS == 0x100D");
    shouldBeTrue("webcl.DEVICE_MAX_READ_IMAGE_ARGS == 0x100E");
    shouldBeTrue("webcl.DEVICE_MAX_WRITE_IMAGE_ARGS == 0x100F");
    shouldBeTrue("webcl.DEVICE_MAX_MEM_ALLOC_SIZE == 0x1010");
    shouldBeTrue("webcl.DEVICE_IMAGE2D_MAX_WIDTH == 0x1011");
    shouldBeTrue("webcl.DEVICE_IMAGE2D_MAX_HEIGHT == 0x1012");
    shouldBeTrue("webcl.DEVICE_IMAGE_SUPPORT == 0x1016");
    shouldBeTrue("webcl.DEVICE_MAX_PARAMETER_SIZE == 0x1017");
    shouldBeTrue("webcl.DEVICE_MAX_SAMPLERS == 0x1018");
    shouldBeTrue("webcl.DEVICE_MEM_BASE_ADDR_ALIGN == 0x1019");
    shouldBeTrue("webcl.DEVICE_SINGLE_FP_CONFIG == 0x101B");
    shouldBeTrue("webcl.DEVICE_GLOBAL_MEM_CACHE_TYPE == 0x101C");
    shouldBeTrue("webcl.DEVICE_GLOBAL_MEM_CACHELINE_SIZE == 0x101D");
    shouldBeTrue("webcl.DEVICE_GLOBAL_MEM_CACHE_SIZE == 0x101E");
    shouldBeTrue("webcl.DEVICE_GLOBAL_MEM_SIZE == 0x101F");
    shouldBeTrue("webcl.DEVICE_MAX_CONSTANT_BUFFER_SIZE == 0x1020");
    shouldBeTrue("webcl.DEVICE_MAX_CONSTANT_ARGS == 0x1021");
    shouldBeTrue("webcl.DEVICE_LOCAL_MEM_TYPE == 0x1022");
    shouldBeTrue("webcl.DEVICE_LOCAL_MEM_SIZE == 0x1023");
    shouldBeTrue("webcl.DEVICE_ERROR_CORRECTION_SUPPORT == 0x1024");
    shouldBeTrue("webcl.DEVICE_PROFILING_TIMER_RESOLUTION == 0x1025");
    shouldBeTrue("webcl.DEVICE_ENDIAN_LITTLE == 0x1026");
    shouldBeTrue("webcl.DEVICE_AVAILABLE == 0x1027");
    shouldBeTrue("webcl.DEVICE_COMPILER_AVAILABLE == 0x1028");
    shouldBeTrue("webcl.DEVICE_EXECUTION_CAPABILITIES == 0x1029");
    shouldBeTrue("webcl.DEVICE_QUEUE_PROPERTIES == 0x102A");
    shouldBeTrue("webcl.DEVICE_PROFILE == 0x102E");
    shouldBeTrue("webcl.DEVICE_VERSION == 0x102F");
    shouldBeTrue("webcl.DEVICE_PLATFORM == 0x1031");
    shouldBeTrue("webcl.DEVICE_HOST_UNIFIED_MEMORY == 0x1035");
    shouldBeTrue("webcl.DEVICE_OPENCL_C_VERSION == 0x103D");

    shouldBeTrue("webcl.FP_DENORM == 0x1");
    shouldBeTrue("webcl.FP_INF_NAN == 0x2");
    shouldBeTrue("webcl.FP_ROUND_TO_NEAREST == 0x4");
    shouldBeTrue("webcl.FP_ROUND_TO_ZERO == 0x8");
    shouldBeTrue("webcl.FP_ROUND_TO_INF == 0x10");
    shouldBeTrue("webcl.FP_FMA == 0x20");
    shouldBeTrue("webcl.FP_SOFT_FLOAT == 0x40");

    shouldBeTrue("webcl.NONE == 0x0");
    shouldBeTrue("webcl.READ_ONLY_CACHE == 0x1");
    shouldBeTrue("webcl.READ_WRITE_CACHE == 0x2");

    shouldBeTrue("webcl.LOCAL == 0x1");
    shouldBeTrue("webcl.GLOBAL == 0x2");

    shouldBeTrue("webcl.EXEC_KERNEL == 0x1");

    shouldBeTrue("webcl.QUEUE_OUT_OF_ORDER_EXEC_MODE_ENABLE == 0x1");
    shouldBeTrue("webcl.QUEUE_PROFILING_ENABLE == 0x2");

    shouldBeTrue("webcl.CONTEXT_DEVICES == 0x1081");
    shouldBeTrue("webcl.CONTEXT_PROPERTIES == 0x1082");
    shouldBeTrue("webcl.CONTEXT_NUM_DEVICES == 0x1083");

    shouldBeTrue("webcl.CONTEXT_PLATFORM == 0x1084");

    shouldBeTrue("webcl.QUEUE_CONTEXT == 0x1090");
    shouldBeTrue("webcl.QUEUE_DEVICE == 0x1091");
    shouldBeTrue("webcl.QUEUE_PROPERTIES == 0x1093");

    shouldBeTrue("webcl.MEM_READ_WRITE == 0x1");
    shouldBeTrue("webcl.MEM_WRITE_ONLY == 0x2");
    shouldBeTrue("webcl.MEM_READ_ONLY == 0x4");

    shouldBeTrue("webcl.R == 0x10B0");
    shouldBeTrue("webcl.A == 0x10B1");
    shouldBeTrue("webcl.RG == 0x10B2");
    shouldBeTrue("webcl.RA == 0x10B3");
    shouldBeTrue("webcl.RGB == 0x10B4");
    shouldBeTrue("webcl.RGBA == 0x10B5");
    shouldBeTrue("webcl.BGRA == 0x10B6");
    shouldBeTrue("webcl.ARGB == 0x10B7");
    shouldBeTrue("webcl.INTENSITY == 0x10B8");
    shouldBeTrue("webcl.LUMINANCE == 0x10B9");
    shouldBeTrue("webcl.Rx == 0x10BA");
    shouldBeTrue("webcl.RGx == 0x10BB");
    shouldBeTrue("webcl.RGBx == 0x10BC");

    shouldBeTrue("webcl.SNORM_INT8 == 0x10D0");
    shouldBeTrue("webcl.SNORM_INT16 == 0x10D1");
    shouldBeTrue("webcl.UNORM_INT8 == 0x10D2");
    shouldBeTrue("webcl.UNORM_INT16 == 0x10D3");
    shouldBeTrue("webcl.UNORM_SHORT_565 == 0x10D4");
    shouldBeTrue("webcl.UNORM_SHORT_555 == 0x10D5");
    shouldBeTrue("webcl.UNORM_INT_101010 == 0x10D6");
    shouldBeTrue("webcl.SIGNED_INT8 == 0x10D7");
    shouldBeTrue("webcl.SIGNED_INT16 == 0x10D8");
    shouldBeTrue("webcl.SIGNED_INT32 == 0x10D9");
    shouldBeTrue("webcl.UNSIGNED_INT8 == 0x10DA");
    shouldBeTrue("webcl.UNSIGNED_INT16 == 0x10DB");
    shouldBeTrue("webcl.UNSIGNED_INT32 == 0x10DC");
    shouldBeTrue("webcl.HALF_FLOAT == 0x10DD");
    shouldBeTrue("webcl.FLOAT == 0x10DE");

    shouldBeTrue("webcl.MEM_OBJECT_BUFFER == 0x10F0");
    shouldBeTrue("webcl.MEM_OBJECT_IMAGE2D == 0x10F1");

    shouldBeTrue("webcl.MEM_TYPE == 0x1100");
    shouldBeTrue("webcl.MEM_FLAGS == 0x1101");
    shouldBeTrue("webcl.MEM_SIZE == 0x1102");

    shouldBeTrue("webcl.MEM_CONTEXT == 0x1106");
    shouldBeTrue("webcl.MEM_ASSOCIATED_MEMOBJECT == 0x1107");
    shouldBeTrue("webcl.MEM_OFFSET == 0x1108");

    shouldBeTrue("webcl.IMAGE_FORMAT == 0x1110");
    shouldBeTrue("webcl.IMAGE_ELEMENT_SIZE == 0x1111");
    shouldBeTrue("webcl.IMAGE_ROW_PITCH == 0x1112");
    shouldBeTrue("webcl.IMAGE_WIDTH == 0x1114");
    shouldBeTrue("webcl.IMAGE_HEIGHT == 0x1115");

    shouldBeTrue("webcl.ADDRESS_NONE == 0x1130");
    shouldBeTrue("webcl.ADDRESS_CLAMP_TO_EDGE == 0x1131");
    shouldBeTrue("webcl.ADDRESS_CLAMP == 0x1132");
    shouldBeTrue("webcl.ADDRESS_REPEAT == 0x1133");
    shouldBeTrue("webcl.ADDRESS_MIRRORED_REPEAT == 0x1134");

    shouldBeTrue("webcl.FILTER_NEAREST == 0x1140");
    shouldBeTrue("webcl.FILTER_LINEAR == 0x1141");

    shouldBeTrue("webcl.SAMPLER_CONTEXT == 0x1151");
    shouldBeTrue("webcl.SAMPLER_NORMALIZED_COORDS == 0x1152");
    shouldBeTrue("webcl.SAMPLER_ADDRESSING_MODE == 0x1153");
    shouldBeTrue("webcl.SAMPLER_FILTER_MODE == 0x1154");

    shouldBeTrue("webcl.MAP_READ == 0x1");
    shouldBeTrue("webcl.MAP_WRITE == 0x2");

    shouldBeTrue("webcl.PROGRAM_CONTEXT == 0x1161");
    shouldBeTrue("webcl.PROGRAM_NUM_DEVICES == 0x1162");
    shouldBeTrue("webcl.PROGRAM_DEVICES == 0x1163");
    shouldBeTrue("webcl.PROGRAM_SOURCE == 0x1164");

    shouldBeTrue("webcl.PROGRAM_BUILD_STATUS == 0x1181");
    shouldBeTrue("webcl.PROGRAM_BUILD_OPTIONS == 0x1182");
    shouldBeTrue("webcl.PROGRAM_BUILD_LOG == 0x1183");

    shouldBeTrue("webcl.BUILD_SUCCESS == 0");
    shouldBeTrue("webcl.BUILD_NONE == -1");
    shouldBeTrue("webcl.BUILD_ERROR == -2");
    shouldBeTrue("webcl.BUILD_IN_PROGRESS == -3");

    shouldBeTrue("webcl.KERNEL_FUNCTION_NAME == 0x1190");
    shouldBeTrue("webcl.KERNEL_NUM_ARGS == 0x1191");
    shouldBeTrue("webcl.KERNEL_CONTEXT == 0x1193");
    shouldBeTrue("webcl.KERNEL_PROGRAM == 0x1194");

    shouldBeTrue("webcl.KERNEL_WORK_GROUP_SIZE == 0x11B0");
    shouldBeTrue("webcl.KERNEL_COMPILE_WORK_GROUP_SIZE == 0x11B1");
    shouldBeTrue("webcl.KERNEL_LOCAL_MEM_SIZE == 0x11B2");
    shouldBeTrue("webcl.KERNEL_PREFERRED_WORK_GROUP_SIZE_MULTIPLE == 0x11B3");
    shouldBeTrue("webcl.KERNEL_PRIVATE_MEM_SIZE == 0x11B4");

    shouldBeTrue("webcl.EVENT_COMMAND_QUEUE == 0x11D0");
    shouldBeTrue("webcl.EVENT_COMMAND_TYPE == 0x11D1");
    shouldBeTrue("webcl.EVENT_COMMAND_EXECUTION_STATUS == 0x11D3");
    shouldBeTrue("webcl.EVENT_CONTEXT == 0x11D4");

    shouldBeTrue("webcl.COMMAND_NDRANGE_KERNEL == 0x11F0");
    shouldBeTrue("webcl.COMMAND_TASK == 0x11F1");
    shouldBeTrue("webcl.COMMAND_READ_BUFFER == 0x11F3");
    shouldBeTrue("webcl.COMMAND_WRITE_BUFFER == 0x11F4");
    shouldBeTrue("webcl.COMMAND_COPY_BUFFER == 0x11F5");
    shouldBeTrue("webcl.COMMAND_READ_IMAGE == 0x11F6");
    shouldBeTrue("webcl.COMMAND_WRITE_IMAGE == 0x11F7");
    shouldBeTrue("webcl.COMMAND_COPY_IMAGE == 0x11F8");
    shouldBeTrue("webcl.COMMAND_COPY_IMAGE_TO_BUFFER == 0x11F9");
    shouldBeTrue("webcl.COMMAND_COPY_BUFFER_TO_IMAGE == 0x11FA");
    shouldBeTrue("webcl.COMMAND_MARKER == 0x11FE");
    shouldBeTrue("webcl.COMMAND_READ_BUFFER_RECT == 0x1201");
    shouldBeTrue("webcl.COMMAND_WRITE_BUFFER_RECT == 0x1202");
    shouldBeTrue("webcl.COMMAND_COPY_BUFFER_RECT == 0x1203");
    shouldBeTrue("webcl.COMMAND_USER == 0x1204");

    shouldBeTrue("webcl.COMPLETE == 0x0");
    shouldBeTrue("webcl.RUNNING == 0x1");
    shouldBeTrue("webcl.SUBMITTED == 0x2");
    shouldBeTrue("webcl.QUEUED == 0x3");

    shouldBeTrue("webcl.PROFILING_COMMAND_QUEUED == 0x1280");
    shouldBeTrue("webcl.PROFILING_COMMAND_SUBMIT == 0x1281");
    shouldBeTrue("webcl.PROFILING_COMMAND_START == 0x1282");
    shouldBeTrue("webcl.PROFILING_COMMAND_END == 0x1283");

    shouldBeTrue("WebCLGL.GL_OBJECT_BUFFER == 0x2000");
    shouldBeTrue("WebCLGL.GL_OBJECT_TEXTURE2D == 0x2001");
    shouldBeTrue("WebCLGL.GL_OBJECT_RENDERBUFFER == 0x2003");
    shouldBeTrue("WebCLGL.GL_TEXTURE_TARGET == 0x2004");
    shouldBeTrue("WebCLGL.GL_MIPMAP_LEVEL == 0x2005");
    shouldBeTrue("WebCLGL.INVALID_GL_OBJECT == -60");
    shouldBeTrue("WebCLGL.INVALID_MIP_LEVEL == -62");
    shouldBeTrue("WebCLGL.COMMAND_ACQUIRE_GL_OBJECTS == 0x11FF");
    shouldBeTrue("WebCLGL.COMMAND_RELEASE_GL_OBJECTS == 0x1200");

}
</script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
        description(document.title);
        debug("Test all constants defined in webcl.idl");
        runTest();
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
