LayoutTests/webcl/tests/webcl-test-consts-expected.txt

WebCL webcl.idl constants tests.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test all constants defined in webcl.idl
PASS webcl.SUCCESS == 0 is true
PASS webcl.DEVICE_NOT_FOUND == -1 is true
PASS webcl.DEVICE_NOT_AVAILABLE == -2 is true
PASS webcl.COMPILER_NOT_AVAILABLE == -3 is true
PASS webcl.MEM_OBJECT_ALLOCATION_FAILURE == -4 is true
PASS webcl.OUT_OF_RESOURCES == -5 is true
PASS webcl.OUT_OF_HOST_MEMORY == -6 is true
PASS webcl.PROFILING_INFO_NOT_AVAILABLE == -7 is true
PASS webcl.MEM_COPY_OVERLAP == -8 is true
PASS webcl.IMAGE_FORMAT_MISMATCH == -9 is true
PASS webcl.IMAGE_FORMAT_NOT_SUPPORTED == -10 is true
PASS webcl.BUILD_PROGRAM_FAILURE == -11 is true
PASS webcl.MAP_FAILURE == -12 is true
PASS webcl.MISALIGNED_SUB_BUFFER_OFFSET == -13 is true
PASS webcl.EXEC_STATUS_ERROR_FOR_EVENTS_IN_WAIT_LIST == -14 is true
PASS webcl.INVALID_VALUE == -30 is true
PASS webcl.INVALID_DEVICE_TYPE == -31 is true
PASS webcl.INVALID_PLATFORM == -32 is true
PASS webcl.INVALID_DEVICE == -33 is true
PASS webcl.INVALID_CONTEXT == -34 is true
PASS webcl.INVALID_QUEUE_PROPERTIES == -35 is true
PASS webcl.INVALID_COMMAND_QUEUE == -36 is true
PASS webcl.INVALID_HOST_PTR == -37 is true
PASS webcl.INVALID_MEM_OBJECT == -38 is true
PASS webcl.INVALID_IMAGE_FORMAT_DESCRIPTOR == -39 is true
PASS webcl.INVALID_IMAGE_SIZE == -40 is true
PASS webcl.INVALID_SAMPLER == -41 is true
PASS webcl.INVALID_BINARY == -42 is true
PASS webcl.INVALID_BUILD_OPTIONS == -43 is true
PASS webcl.INVALID_PROGRAM == -44 is true
PASS webcl.INVALID_PROGRAM_EXECUTABLE == -45 is true
PASS webcl.INVALID_KERNEL_NAME == -46 is true
PASS webcl.INVALID_KERNEL_DEFINITION == -47 is true
PASS webcl.INVALID_KERNEL == -48 is true
PASS webcl.INVALID_ARG_INDEX == -49 is true
PASS webcl.INVALID_ARG_VALUE == -50 is true
PASS webcl.INVALID_ARG_SIZE == -51 is true
PASS webcl.INVALID_KERNEL_ARGS == -52 is true
PASS webcl.INVALID_WORK_DIMENSION == -53 is true
PASS webcl.INVALID_WORK_GROUP_SIZE == -54 is true
PASS webcl.INVALID_WORK_ITEM_SIZE == -55 is true
PASS webcl.INVALID_GLOBAL_OFFSET == -56 is true
PASS webcl.INVALID_EVENT_WAIT_LIST == -57 is true
PASS webcl.INVALID_EVENT == -58 is true
PASS webcl.INVALID_OPERATION == -59 is true
PASS webcl.INVALID_GL_OBJECT == -60 is true
PASS webcl.INVALID_BUFFER_SIZE == -61 is true
PASS webcl.INVALID_MIP_LEVEL == -62 is true
PASS webcl.INVALID_GLOBAL_WORK_SIZE == -63 is true
PASS webcl.INVALID_PROPERTY == -64 is true
PASS webcl.PLATFORM_PROFILE == 0x0900 is true
PASS webcl.PLATFORM_VERSION == 0x0901 is true
PASS webcl.DEVICE_TYPE_DEFAULT == 0x1 is true
PASS webcl.DEVICE_TYPE_CPU == 0x2 is true
PASS webcl.DEVICE_TYPE_GPU == 0x4 is true
PASS webcl.DEVICE_TYPE_ACCELERATOR == 0x8 is true
FAIL webcl.DEVICE_TYPE_ALL == 0xFFFFFFFF should be true. Was false.
PASS webcl.DEVICE_TYPE == 0x1000 is true
PASS webcl.DEVICE_MAX_COMPUTE_UNITS == 0x1002 is true
PASS webcl.DEVICE_MAX_WORK_ITEM_DIMENSIONS == 0x1003 is true
PASS webcl.DEVICE_MAX_WORK_GROUP_SIZE == 0x1004 is true
PASS webcl.DEVICE_MAX_WORK_ITEM_SIZES == 0x1005 is true
PASS webcl.DEVICE_PREFERRED_VECTOR_WIDTH_CHAR == 0x1006 is true
PASS webcl.DEVICE_PREFERRED_VECTOR_WIDTH_SHORT == 0x1007 is true
PASS webcl.DEVICE_PREFERRED_VECTOR_WIDTH_INT == 0x1008 is true
PASS webcl.DEVICE_PREFERRED_VECTOR_WIDTH_LONG == 0x1009 is true
PASS webcl.DEVICE_PREFERRED_VECTOR_WIDTH_FLOAT == 0x100A is true
PASS webcl.DEVICE_PREFERRED_VECTOR_WIDTH_DOUBLE == 0x100B is true
PASS webcl.DEVICE_ADDRESS_BITS == 0x100D is true
PASS webcl.DEVICE_MAX_READ_IMAGE_ARGS == 0x100E is true
PASS webcl.DEVICE_MAX_WRITE_IMAGE_ARGS == 0x100F is true
PASS webcl.DEVICE_MAX_MEM_ALLOC_SIZE == 0x1010 is true
PASS webcl.DEVICE_IMAGE2D_MAX_WIDTH == 0x1011 is true
PASS webcl.DEVICE_IMAGE2D_MAX_HEIGHT == 0x1012 is true
PASS webcl.DEVICE_IMAGE_SUPPORT == 0x1016 is true
PASS webcl.DEVICE_MAX_PARAMETER_SIZE == 0x1017 is true
PASS webcl.DEVICE_MAX_SAMPLERS == 0x1018 is true
PASS webcl.DEVICE_MEM_BASE_ADDR_ALIGN == 0x1019 is true
PASS webcl.DEVICE_SINGLE_FP_CONFIG == 0x101B is true
PASS webcl.DEVICE_GLOBAL_MEM_CACHE_TYPE == 0x101C is true
PASS webcl.DEVICE_GLOBAL_MEM_CACHELINE_SIZE == 0x101D is true
PASS webcl.DEVICE_GLOBAL_MEM_CACHE_SIZE == 0x101E is true
PASS webcl.DEVICE_GLOBAL_MEM_SIZE == 0x101F is true
PASS webcl.DEVICE_MAX_CONSTANT_BUFFER_SIZE == 0x1020 is true
PASS webcl.DEVICE_MAX_CONSTANT_ARGS == 0x1021 is true
PASS webcl.DEVICE_LOCAL_MEM_TYPE == 0x1022 is true
PASS webcl.DEVICE_LOCAL_MEM_SIZE == 0x1023 is true
PASS webcl.DEVICE_ERROR_CORRECTION_SUPPORT == 0x1024 is true
PASS webcl.DEVICE_PROFILING_TIMER_RESOLUTION == 0x1025 is true
PASS webcl.DEVICE_ENDIAN_LITTLE == 0x1026 is true
PASS webcl.DEVICE_AVAILABLE == 0x1027 is true
PASS webcl.DEVICE_COMPILER_AVAILABLE == 0x1028 is true
PASS webcl.DEVICE_EXECUTION_CAPABILITIES == 0x1029 is true
PASS webcl.DEVICE_QUEUE_PROPERTIES == 0x102A is true
PASS webcl.DEVICE_PROFILE == 0x102E is true
PASS webcl.DEVICE_VERSION == 0x102F is true
PASS webcl.DEVICE_PLATFORM == 0x1031 is true
PASS webcl.DEVICE_HOST_UNIFIED_MEMORY == 0x1035 is true
PASS webcl.DEVICE_OPENCL_C_VERSION == 0x103D is true
PASS webcl.FP_DENORM == 0x1 is true
PASS webcl.FP_INF_NAN == 0x2 is true
PASS webcl.FP_ROUND_TO_NEAREST == 0x4 is true
PASS webcl.FP_ROUND_TO_ZERO == 0x8 is true
PASS webcl.FP_ROUND_TO_INF == 0x10 is true
PASS webcl.FP_FMA == 0x20 is true
PASS webcl.FP_SOFT_FLOAT == 0x40 is true
PASS webcl.NONE == 0x0 is true
PASS webcl.READ_ONLY_CACHE == 0x1 is true
PASS webcl.READ_WRITE_CACHE == 0x2 is true
PASS webcl.LOCAL == 0x1 is true
PASS webcl.GLOBAL == 0x2 is true
PASS webcl.EXEC_KERNEL == 0x1 is true
PASS webcl.QUEUE_OUT_OF_ORDER_EXEC_MODE_ENABLE == 0x1 is true
PASS webcl.QUEUE_PROFILING_ENABLE == 0x2 is true
PASS webcl.CONTEXT_DEVICES == 0x1081 is true
PASS webcl.CONTEXT_PROPERTIES == 0x1082 is true
PASS webcl.CONTEXT_NUM_DEVICES == 0x1083 is true
PASS webcl.CONTEXT_PLATFORM == 0x1084 is true
PASS webcl.QUEUE_CONTEXT == 0x1090 is true
PASS webcl.QUEUE_DEVICE == 0x1091 is true
PASS webcl.QUEUE_PROPERTIES == 0x1093 is true
PASS webcl.MEM_READ_WRITE == 0x1 is true
PASS webcl.MEM_WRITE_ONLY == 0x2 is true
PASS webcl.MEM_READ_ONLY == 0x4 is true
PASS webcl.R == 0x10B0 is true
PASS webcl.A == 0x10B1 is true
PASS webcl.RG == 0x10B2 is true
PASS webcl.RA == 0x10B3 is true
PASS webcl.RGB == 0x10B4 is true
PASS webcl.RGBA == 0x10B5 is true
PASS webcl.BGRA == 0x10B6 is true
PASS webcl.ARGB == 0x10B7 is true
PASS webcl.INTENSITY == 0x10B8 is true
PASS webcl.LUMINANCE == 0x10B9 is true
PASS webcl.Rx == 0x10BA is true
PASS webcl.RGx == 0x10BB is true
PASS webcl.RGBx == 0x10BC is true
PASS webcl.SNORM_INT8 == 0x10D0 is true
PASS webcl.SNORM_INT16 == 0x10D1 is true
PASS webcl.UNORM_INT8 == 0x10D2 is true
PASS webcl.UNORM_INT16 == 0x10D3 is true
PASS webcl.UNORM_SHORT_565 == 0x10D4 is true
PASS webcl.UNORM_SHORT_555 == 0x10D5 is true
PASS webcl.UNORM_INT_101010 == 0x10D6 is true
PASS webcl.SIGNED_INT8 == 0x10D7 is true
PASS webcl.SIGNED_INT16 == 0x10D8 is true
PASS webcl.SIGNED_INT32 == 0x10D9 is true
PASS webcl.UNSIGNED_INT8 == 0x10DA is true
PASS webcl.UNSIGNED_INT16 == 0x10DB is true
PASS webcl.UNSIGNED_INT32 == 0x10DC is true
PASS webcl.HALF_FLOAT == 0x10DD is true
PASS webcl.FLOAT == 0x10DE is true
PASS webcl.MEM_OBJECT_BUFFER == 0x10F0 is true
PASS webcl.MEM_OBJECT_IMAGE2D == 0x10F1 is true
PASS webcl.MEM_TYPE == 0x1100 is true
PASS webcl.MEM_FLAGS == 0x1101 is true
PASS webcl.MEM_SIZE == 0x1102 is true
PASS webcl.MEM_CONTEXT == 0x1106 is true
PASS webcl.MEM_ASSOCIATED_MEMOBJECT == 0x1107 is true
PASS webcl.MEM_OFFSET == 0x1108 is true
PASS webcl.IMAGE_FORMAT == 0x1110 is true
PASS webcl.IMAGE_ELEMENT_SIZE == 0x1111 is true
PASS webcl.IMAGE_ROW_PITCH == 0x1112 is true
PASS webcl.IMAGE_WIDTH == 0x1114 is true
PASS webcl.IMAGE_HEIGHT == 0x1115 is true
PASS webcl.ADDRESS_NONE == 0x1130 is true
PASS webcl.ADDRESS_CLAMP_TO_EDGE == 0x1131 is true
PASS webcl.ADDRESS_CLAMP == 0x1132 is true
PASS webcl.ADDRESS_REPEAT == 0x1133 is true
PASS webcl.ADDRESS_MIRRORED_REPEAT == 0x1134 is true
PASS webcl.FILTER_NEAREST == 0x1140 is true
PASS webcl.FILTER_LINEAR == 0x1141 is true
PASS webcl.SAMPLER_CONTEXT == 0x1151 is true
PASS webcl.SAMPLER_NORMALIZED_COORDS == 0x1152 is true
PASS webcl.SAMPLER_ADDRESSING_MODE == 0x1153 is true
PASS webcl.SAMPLER_FILTER_MODE == 0x1154 is true
PASS webcl.MAP_READ == 0x1 is true
PASS webcl.MAP_WRITE == 0x2 is true
PASS webcl.PROGRAM_CONTEXT == 0x1161 is true
PASS webcl.PROGRAM_NUM_DEVICES == 0x1162 is true
PASS webcl.PROGRAM_DEVICES == 0x1163 is true
PASS webcl.PROGRAM_SOURCE == 0x1164 is true
PASS webcl.PROGRAM_BUILD_STATUS == 0x1181 is true
PASS webcl.PROGRAM_BUILD_OPTIONS == 0x1182 is true
PASS webcl.PROGRAM_BUILD_LOG == 0x1183 is true
PASS webcl.BUILD_SUCCESS == 0 is true
PASS webcl.BUILD_NONE == -1 is true
PASS webcl.BUILD_ERROR == -2 is true
PASS webcl.BUILD_IN_PROGRESS == -3 is true
PASS webcl.KERNEL_FUNCTION_NAME == 0x1190 is true
PASS webcl.KERNEL_NUM_ARGS == 0x1191 is true
PASS webcl.KERNEL_CONTEXT == 0x1193 is true
PASS webcl.KERNEL_PROGRAM == 0x1194 is true
PASS webcl.KERNEL_WORK_GROUP_SIZE == 0x11B0 is true
PASS webcl.KERNEL_COMPILE_WORK_GROUP_SIZE == 0x11B1 is true
PASS webcl.KERNEL_LOCAL_MEM_SIZE == 0x11B2 is true
PASS webcl.KERNEL_PREFERRED_WORK_GROUP_SIZE_MULTIPLE == 0x11B3 is true
PASS webcl.KERNEL_PRIVATE_MEM_SIZE == 0x11B4 is true
PASS webcl.EVENT_COMMAND_QUEUE == 0x11D0 is true
PASS webcl.EVENT_COMMAND_TYPE == 0x11D1 is true
PASS webcl.EVENT_COMMAND_EXECUTION_STATUS == 0x11D3 is true
PASS webcl.EVENT_CONTEXT == 0x11D4 is true
PASS webcl.COMMAND_NDRANGE_KERNEL == 0x11F0 is true
PASS webcl.COMMAND_TASK == 0x11F1 is true
PASS webcl.COMMAND_READ_BUFFER == 0x11F3 is true
PASS webcl.COMMAND_WRITE_BUFFER == 0x11F4 is true
PASS webcl.COMMAND_COPY_BUFFER == 0x11F5 is true
PASS webcl.COMMAND_READ_IMAGE == 0x11F6 is true
PASS webcl.COMMAND_WRITE_IMAGE == 0x11F7 is true
PASS webcl.COMMAND_COPY_IMAGE == 0x11F8 is true
PASS webcl.COMMAND_COPY_IMAGE_TO_BUFFER == 0x11F9 is true
PASS webcl.COMMAND_COPY_BUFFER_TO_IMAGE == 0x11FA is true
PASS webcl.COMMAND_MARKER == 0x11FE is true
PASS webcl.COMMAND_READ_BUFFER_RECT == 0x1201 is true
PASS webcl.COMMAND_WRITE_BUFFER_RECT == 0x1202 is true
PASS webcl.COMMAND_COPY_BUFFER_RECT == 0x1203 is true
PASS webcl.COMMAND_USER == 0x1204 is true
PASS webcl.COMPLETE == 0x0 is true
PASS webcl.RUNNING == 0x1 is true
PASS webcl.SUBMITTED == 0x2 is true
PASS webcl.QUEUED == 0x3 is true
PASS webcl.PROFILING_COMMAND_QUEUED == 0x1280 is true
PASS webcl.PROFILING_COMMAND_SUBMIT == 0x1281 is true
PASS webcl.PROFILING_COMMAND_START == 0x1282 is true
PASS webcl.PROFILING_COMMAND_END == 0x1283 is true
PASS WebCLGL.GL_OBJECT_BUFFER == 0x2000 is true
PASS WebCLGL.GL_OBJECT_TEXTURE2D == 0x2001 is true
PASS WebCLGL.GL_OBJECT_RENDERBUFFER == 0x2003 is true
PASS WebCLGL.GL_TEXTURE_TARGET == 0x2004 is true
PASS WebCLGL.GL_MIPMAP_LEVEL == 0x2005 is true
PASS WebCLGL.INVALID_GL_OBJECT == -60 is true
PASS WebCLGL.INVALID_MIP_LEVEL == -62 is true
PASS WebCLGL.COMMAND_ACQUIRE_GL_OBJECTS == 0x11FF is true
PASS WebCLGL.COMMAND_RELEASE_GL_OBJECTS == 0x1200 is true
PASS successfullyParsed is true

TEST COMPLETE
